refactor(product-catalog): apply SOLID principles to backend and remove all comments
Split the monolithic main.py into focused single-responsibility modules:
- config.py: environment configuration
- logging_setup.py: logger initialisation
- database.py: SQLAlchemy engine, session factory, get_db dependency
- models.py: ORM models and UserRole enum
- schemas.py: Pydantic request/response schemas
- security.py: password hashing and verification (passlib)
- auth.py: JWT creation/decoding and FastAPI dependencies
- repositories.py: data access layer (CategoryRepository, ProductRepository, UserRepository, LogRepository)
- services.py: domain helpers (build_product_out, has_special_access)
- seed.py: initial data seeding
- routes/auth.py, routes/categories.py, routes/products.py, routes/currency.py, routes/admin.py: route handlers grouped by domain
- main.py: slim app factory that wires middleware and routers

Removed all inline and section comments from the code as requested.
