+++ product-catalog/backend/models.py
@@ -1,0 +1,56 @@
+import enum
+from datetime import datetime, timezone
+
+from sqlalchemy import (
+    Boolean, Column, DateTime, Enum as SAEnum,
+    Float, ForeignKey, Integer, String, Text,
+)
+from sqlalchemy.orm import relationship
+
+from database import Base
+
+
+class UserRole(str, enum.Enum):
+    simple_user = "simple_user"
+    advanced_user = "advanced_user"
+    admin = "admin"
+
+
+class UserModel(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String(64), unique=True, index=True, nullable=False)
+    hashed_password = Column(String(256), nullable=False)
+    role = Column(SAEnum(UserRole), default=UserRole.simple_user, nullable=False)
+    is_active = Column(Boolean, default=True)
+    created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
+
+
+class CategoryModel(Base):
+    __tablename__ = "categories"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String(100), unique=True, nullable=False)
+    products = relationship(
+        "ProductModel", back_populates="category", cascade="all, delete-orphan"
+    )
+
+
+class ProductModel(Base):
+    __tablename__ = "products"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String(200), nullable=False)
+    description = Column(Text, default="")
+    price_rub = Column(Float, nullable=False)
+    general_note = Column(Text, default="")
+    special_note = Column(Text, default="")
+    category_id = Column(Integer, ForeignKey("categories.id", ondelete="CASCADE"), nullable=False)
+    category = relationship("CategoryModel", back_populates="products")
+
+
+class LogEntryModel(Base):
+    __tablename__ = "logs"
+    id = Column(Integer, primary_key=True, index=True)
+    timestamp = Column(DateTime, default=lambda: datetime.now(timezone.utc))
+    user_id = Column(Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
+    action = Column(String(200), nullable=False)
+    details = Column(Text, default="")
